Moved parseEntry to plugin.utils
